backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pandas==2.2.3
yfinance==0.2.43
simpleeval==0.9.13
python-dotenv==1.0.1
apscheduler==3.10.4
